+++ src/main/java/org/trebol/jpa/entities/SellDetail.java
@@ -33,6 +33,8 @@ public class SellDetail
   private Long id;
   @Column(name = "sell_detail_units", nullable = false)
   private int units;
+  @Column(name = "sell_detail_unit_value", nullable = false)
+  private Integer unitValue;
   @JoinColumn(name = "product_id", referencedColumnName = "product_id", updatable = false)
   @ManyToOne(optional = false)
   private Product product;
@@ -61,6 +63,14 @@ public class SellDetail
     this.units = units;
   }
 
+  public Integer getUnitValue() {
+    return unitValue;
+  }
+
+  public void setUnitValue(Integer unitValue) {
+    this.unitValue = unitValue;
+  }
+
   public Product getProduct() {
     return product;
   }
@@ -70,39 +80,25 @@ public class SellDetail
   }
 
   @Override
-  public int hashCode() {
-    int hash = 3;
-    hash = 59 * hash + Objects.hashCode(this.id);
-    hash = 59 * hash + this.units;
-    hash = 59 * hash + Objects.hashCode(this.product);
-    return hash;
+  public boolean equals(Object o) {
+    if (this == o) return true;
+    if (o == null || getClass() != o.getClass()) return false;
+    SellDetail that = (SellDetail) o;
+    return units == that.units && Objects.equals(id, that.id) && Objects.equals(unitValue, that.unitValue) && Objects.equals(product, that.product);
   }
 
   @Override
-  public boolean equals(Object obj) {
-    if (this == obj) {
-      return true;
-    }
-    if (obj == null) {
-      return false;
-    }
-    if (getClass() != obj.getClass()) {
-      return false;
-    }
-    final SellDetail other = (SellDetail)obj;
-    if (this.units != other.units) {
-      return false;
-    }
-    if (!Objects.equals(this.id, other.id)) {
-      return false;
-    }
-    return Objects.equals(this.product, other.product);
+  public int hashCode() {
+    return Objects.hash(id, units, unitValue, product);
   }
 
   @Override
   public String toString() {
-    return "SellDetail{id=" + id +
+    return "SellDetail{" +
+        "id=" + id +
         ", units=" + units +
-        ", product=" + product + '}';
+        ", unitValue=" + unitValue +
+        ", product=" + product +
+        '}';
   }
 }